move compact command to archiver.compact
